\docType{data}
\name{scotpop19}
\alias{scotpop19}
\title{Mid-year population estimates}
\format{
tabular object
}
\source{
https://statistics.gov.scot/resource?uri=http%3A%2F%2Fstatistics.gov.scot%2Fdata%2Fpopulation-estimates-detailed-current-geographic-http%3A%2F%2Fstatistics.gov.scot%2Fdata%2Fpopulation-estimates-2011-datazone-flatfiles
}
\usage{
scotpop19
}
\description{
Annual publication of mid-year population estimates by age and sex as at 30 June every year, where population for higher geographies is aggregated from the population estimates for 2011 Data Zones.
}
\references{
}
\keyword{datasets}
